api/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
pydantic-settings
mysql-connector-python